pos_backend/requirements.txt
# Django and DRF
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# CORS
django-cors-headers>=4.3.0

# Filtering
django-filter>=23.3

# Database (optional, uncomment if using PostgreSQL)
# psycopg2-binary>=2.9.9

# Utilities
python-decouple>=3.8
Pillow>=10.1.0

# Development tools (optional)
# django-debug-toolbar>=4.2.0
# ipython>=8.17.0
